combine sequence classification for thai2fit
